+++ docs/.vuepress/config.js
@@ -1,7 +1,7 @@
 const path = require('path')
 
 module.exports = {
-  title: 'Kong Design System',
+  title: 'Kongponents',
   description: 'Kong UI Components & Style Guide',
   themeConfig: {
     repo: 'kong/kongponents',
@@ -12,14 +12,11 @@ module.exports = {
     search: false,
     sidebar: [
       {
-        title: 'Style Guide',
+        title: 'Guide',
         collapsable: false,
         children: [
-          '/style-guide/colors',
-          '/style-guide/type',
-          '/style-guide/spacing',
-          '/style-guide/forms',
-          '/style-guide/theming'
+          '/',
+          '/guide/theming'
         ]
       },
       {
@@ -27,7 +24,17 @@ module.exports = {
         collapsable: false,
         sidebarDepth: 1,
         children: [
-          '/components/',
+          {
+            title: 'Style Guide',
+            collapsable: false,
+            children: [
+              '/style-guide/colors',
+              '/style-guide/type',
+              '/style-guide/forms',
+              '/style-guide/utilities',
+
+            ]
+          },
           {
             title: 'Components',
             collapsable: false,
@@ -35,8 +42,10 @@ module.exports = {
             children: [
               '/components/alert',
               '/components/badge',
+              '/components/breadcrumbs',
               '/components/button',
               '/components/card',
+              '/components/empty-state',
               '/components/input',
               '/components/input-checkbox',
               '/components/input-radio',
@@ -45,6 +54,8 @@ module.exports = {
               '/components/popover',
               '/components/tooltip',
               '/components/table',
+              '/components/toaster',
+              '/components/slideout',
               '/components/switch'
             ]
           },
@@ -59,9 +70,13 @@ module.exports = {
           }
         ]
       },
-    ],
-    nav: [
-      { text: 'Home', link: '/' }
+      {
+        title: 'Contributing',
+        collapsable: false,
+        children: [
+          '/contributing/local'
+        ]
+      }
     ]
   },
   head: [
@@ -92,5 +107,8 @@ module.exports = {
       .use('raw')
       .loader('raw-loader')
       .end().end()
-  }
+  },
+  globalUIComponents: [
+    'GlobalUI'
+  ]
 }