+++ docs/.vuepress/enhanceApp.js
@@ -3,25 +3,23 @@ import '../../packages/styles/styles.css'
 import icons from '../../packages/KIcon/icons' // KIcon icons
 
 // Globally import all Kongponents
-import KAlert from '../../packages/KAlert'
-import KBadge from '../../packages/KBadge'
-import KButton from '../../packages/KButton'
-import KCard from '../../packages/KCard'
-import KClipboardProvider from '../../packages/KClipboardProvider'
-import KEmptyState from '../../packages/KEmptyState'
-import KIcon from '../../packages/KIcon'
-import KInput from '../../packages/KInput'
-import KInputSwitch from '../../packages/KInputSwitch'
-import KInputCheckbox from '../../packages/KInputCheckbox'
-import KInputRadio from '../../packages/KInputRadio'
-import KLabel from '../../packages/KLabel'
-import KModal from '../../packages/KModal'
-import KoolTip from '../../packages/KoolTip'
-import KPop from '../../packages/KPop'
-import Krumbs from '../../packages/Krumbs'
-import KTable from '../../packages/KTable'
-import KToaster from '../../packages/KToaster'
-import KToggle from '../../packages/KToggle'
+import KAlert from '../../packages/KAlert/KAlert.vue'
+import KBadge from '../../packages/KBadge/KBadge.vue'
+import KButton from '../../packages/KButton/KButton.vue'
+import KCard from '../../packages/KCard/KCard.vue'
+import KClipboardProvider from '../../packages/KClipboardProvider/KClipboardProvider.js'
+import KEmptyState from '../../packages/KEmptyState/KEmptyState.vue'
+import KIcon from '../../packages/KIcon/KIcon.vue'
+import KInput from '../../packages/KInput/KInput.vue'
+import KInputSwitch from '../../packages/KInputSwitch/KInputSwitch.vue'
+import KLabel from '../../packages/KLabel/KLabel.vue'
+import KModal from '../../packages/KModal/KModal.vue'
+import KoolTip from '../../packages/KoolTip/KoolTip.vue'
+import KPop from '../../packages/KPop/KPop.vue'
+import Krumbs from '../../packages/Krumbs/Krumbs.vue'
+import KTable from '../../packages/KTable/KTable.vue'
+import KToggle from '../../packages/KToggle/KToggle.js'
+import KSlideout from '../../packages/KSlideout/KSlideout.vue'
 
 export default ({
   Vue,
@@ -46,8 +44,8 @@ export default ({
   Vue.component('KPop', KPop)
   Vue.component('Krumbs', Krumbs)
   Vue.component('KTable', KTable)
-  Vue.component('KToaster', KToaster)
   Vue.component('KToggle', KToggle)
+  Vue.component('KSlideout', KSlideout)
 
   Vue.prototype.$icons = Object.keys(icons)
 }