+++ packages/KButton/CHANGELOG.md
@@ -1,3 +1,35 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.1.10-4](https://github.com/Kong/kongponents/compare/@kongponents/kbutton@0.1.10-3...@kongponents/kbutton@0.1.10-4) (2019-11-11)
+
+**Note:** Version bump only for package @kongponents/kbutton
+
+
+
+
+
+## [0.1.10-3](https://github.com/Kong/kongponents/compare/@kongponents/kbutton@0.1.10-2...@kongponents/kbutton@0.1.10-3) (2019-11-08)
+
+
+### Bug Fixes
+
+* **dist:** dist to all packages ([#152](https://github.com/Kong/kongponents/issues/152)) ([217c46b](https://github.com/Kong/kongponents/commit/217c46b02afe960f8fa597ba77410dcf034b1716))
+
+
+
+
+
+## [0.1.10-2](https://github.com/Kong/kongponents/compare/@kongponents/kbutton@0.1.10-1...@kongponents/kbutton@0.1.10-2) (2019-11-08)
+
+**Note:** Version bump only for package @kongponents/kbutton
+
+
+
+
+
 # @kongponents/kbutton
 
 ## Table of Contents