+++ packages/KCard/CHANGELOG.md
@@ -1,9 +1,46 @@
-# @kongponents/kcard
+# Change Log
 
-## Table of Contents
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-- [0.1.3-1](#013-1)
-- [0.1.3](#013)
+## [0.1.3-6](https://github.com/Kong/kongponents/compare/@kongponents/kcard@0.1.3-5...@kongponents/kcard@0.1.3-6) (2019-11-11)
+
+**Note:** Version bump only for package @kongponents/kcard
+
+
+
+
+
+## [0.1.3-5](https://github.com/Kong/kongponents/compare/@kongponents/kcard@0.1.3-4...@kongponents/kcard@0.1.3-5) (2019-11-08)
+
+
+### Bug Fixes
+
+* **dist:** dist to all packages ([#152](https://github.com/Kong/kongponents/issues/152)) ([217c46b](https://github.com/Kong/kongponents/commit/217c46b02afe960f8fa597ba77410dcf034b1716))
+
+
+
+
+
+## [0.1.3-4](https://github.com/Kong/kongponents/compare/@kongponents/kcard@0.1.3-3...@kongponents/kcard@0.1.3-4) (2019-11-08)
+
+**Note:** Version bump only for package @kongponents/kcard
+
+
+
+
+
+## [0.1.3-3](https://github.com/Kong/kongponents/compare/@kongponents/kcard@0.1.3-1...@kongponents/kcard@0.1.3-3) (2019-11-06)
+
+**Note:** Version bump only for package @kongponents/kcard
+
+
+
+
+
+## [0.1.3-2](https://github.com/Kong/kongponents/compare/@kongponents/kcard@0.1.3-1...@kongponents/kcard@0.1.3-2) (2019-11-06)
+
+**Note:** Version bump only for package @kongponents/kcard
 
 ### 0.1.3-1
  - check `title` slot in `$slots` for backwards compat with Vue version < 2.6 [#120](https://github.com/Kong/kongponents/pull/120)