+++ packages/KIcon/CHANGELOG.md
@@ -1,9 +1,45 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.3.4](https://github.com/Kong/kongponents/compare/@kongponents/kicon@0.3.3...@kongponents/kicon@0.3.4) (2019-11-11)
+
+**Note:** Version bump only for package @kongponents/kicon
+
+
+
+
+
+## [0.3.3](https://github.com/Kong/kongponents/compare/@kongponents/kicon@0.3.2...@kongponents/kicon@0.3.3) (2019-11-08)
+
+
+### Bug Fixes
+
+* **dist:** dist to all packages ([#152](https://github.com/Kong/kongponents/issues/152)) ([217c46b](https://github.com/Kong/kongponents/commit/217c46b02afe960f8fa597ba77410dcf034b1716))
+
+
+
+
+
+## [0.3.2](https://github.com/Kong/kongponents/compare/@kongponents/kicon@0.3.1...@kongponents/kicon@0.3.2) (2019-11-08)
+
+**Note:** Version bump only for package @kongponents/kicon
+
+
+
+
+
 # @kongponents/kicon
 
 ## Table of Contents
 
+- [0.3.1](#134)
 - [0.3.0](#129)
 - [0.2.3](#023)
+
+### 0.3.1
+ - Adds two new icons, close (x) and lock [#134](https://github.com/Kong/kongponents/pull/134)
 
 ### 0.3.0
  - Adds two new icons, warning and dangerCircle and uses a blacklisting mechanism to preserve fill colors of symbols inside of an icon [#127](https://github.com/Kong/kongponents/pull/129)